+++ src/Controller/UserController.php
@@ -97,20 +97,24 @@ class UserController extends AbstractController
     public function edit(Request $request, User $user): Response
     {
         if($this->session->get('is_logged_in') == true) {
-            $form = $this->createForm(UserType::class, $user);
-            $form->handleRequest($request);
+            if($this->session->get('user_id') == $user->getId()) {
+                $form = $this->createForm(UserType::class, $user);
+                $form->handleRequest($request);
 
-            if ($form->isSubmitted() && $form->isValid()) {
-                //$this->getDoctrine()->getManager()->flush();
-                $entityManager = $this->getDoctrine()->getManager();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    //$this->getDoctrine()->getManager()->flush();
+                    $entityManager = $this->getDoctrine()->getManager();
 
-                $plainPassword = $user->getPassword();
-                $encoded = $this->passwordEncoder->encodePassword($user, $plainPassword);
-                $user->setPassword($encoded);
+                    $plainPassword = $user->getPassword();
+                    $encoded = $this->passwordEncoder->encodePassword($user, $plainPassword);
+                    $user->setPassword($encoded);
 
-                $entityManager->persist($user);
-                $entityManager->flush();
+                    $entityManager->persist($user);
+                    $entityManager->flush();
 
+                    return $this->redirectToRoute('user_index');
+                }
+            } else {
                 return $this->redirectToRoute('user_index');
             }
 
@@ -129,13 +133,13 @@ class UserController extends AbstractController
     public function delete(Request $request, User $user): Response
     {
         if($this->session->get('is_logged_in') == true) {
-            if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token')))
-            {
-                $entityManager = $this->getDoctrine()->getManager();
-                $entityManager->remove($user);
-                $entityManager->flush();
+            if($this->session->get('user_id') == $user->getId()) {
+                if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($user);
+                    $entityManager->flush();
+                }
             }
-
             return $this->redirectToRoute('user_index');
         } else {
             return $this->redirectToRoute('app_login');
